+++ build.gradle
@@ -23,6 +23,7 @@ targetCompatibility = JavaVersion.VERSION_11
 repositories {
     mavenCentral()
     maven { url 'https://oss.sonatype.org/content/repositories/snapshots/' }
+    maven { url 'https://mvnrepository.com/artifact/com.itextpdf/kernel/7.0.4'}
 }
 
 checkstyle {
@@ -61,10 +62,11 @@ dependencies {
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.7.0'
     implementation group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-jsr310', version: '2.7.4'
 
+    implementation group: 'com.itextpdf', name: 'kernel', verision: '7.0.4'
+
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: jUnitVersion
 
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: jUnitVersion
-    compile fileTree(dir:'libs',include:['*.jar'])
 }
 
 shadowJar {
